Switch PTVSD over to Gunicorn
Fixes #998
