+++ manage.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 import os, sys, warnings, logging
 
+from django.core.management import execute_from_command_line
+
 logger = logging.getLogger(__name__)
 
 if __name__ == "__main__":
@@ -10,7 +12,5 @@ if __name__ == "__main__":
     # Regex for which modules to ignore warnings from
     IGNORE_MODULES = 'djangosaml2'
     warnings.filterwarnings("ignore", module=IGNORE_MODULES, category=DeprecationWarning)
-
-    from django.core.management import execute_from_command_line
 
     execute_from_command_line(sys.argv)
