+++ requirements.txt
@@ -43,7 +43,6 @@ google-cloud-bigquery==1.26.0
 
 pinax-eventlog==2.0.3
 ptvsd==4.3.2
-django-ptvsd-debug==1.0.3
 PyLTI1p3==1.7.0
 django-webpack-loader==0.6.0
 jsonschema==3.2.0
